Добавлен тест для проверки метода конвертации

+++ console/migrations/m130524_201442_init.php
@@ -19,7 +19,7 @@ class m130524_201442_init extends Migration
             'password_hash' => $this->string()->notNull(),
             'password_reset_token' => $this->string()->unique(),
             'email' => $this->string()->notNull()->unique(),
-            'bonus' => $this->float()->defaultValue(0)->notNull()->comment('Баллы пользователя'),
+            'bonus' => $this->float(2)->defaultValue(0)->notNull()->comment('Баллы пользователя'),
             'status' => $this->smallInteger()->notNull()->defaultValue(10),
             'created_at' => $this->integer()->notNull(),
             'updated_at' => $this->integer()->notNull(),
